scanner_requirements.txt
certifi==2020.4.5.2
chardet==3.0.4
idna==2.9
pika==1.1.0
psycopg2-binary==2.8.5
Pypubsub==4.0.3
python-binance-chain==0.1.20
requests==2.23.0
urllib3==1.25.9
web3==5.11.1
two1==3.10.9
secp256k1==0.13.2
